fix: protect against concurrent cache access
There are two threads using the fetchedRecords collection - the
Kafka Connect thread that removes items from the collection when
it calls the task's poll() method, and the TimerTask thread that
is adding items to the collection.

This commit protects against a concurrent access exception that
can occur if both of those threads trigger at the same time.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>